MINOR: examples: expand rate limit rule examples
We recently released an update to the rate limit rules user experience.
While this change did not impact the API it did introduce a new signal
`ALL-REQUESTS` that can be used when configuring rate limit rules.
It has also been requested that we expand the terraform examples to show
how different configurations that are done through the UI could be done
within terraform.

This commit adds all 3 examples that could be configured from within
the UI.

The example does not include the corresponding signal creation as the
existing rule also did not include that.