Add responseHeader to valid condition fields (#168)
* Add responseHeader to the valid condition fields

According to Fastly support and the SigSci dashboard, response headers
are a valid attribute to set conditions against. The TF provider does
not currently agree and the spurious warnings make it hard to debug
other SigSci issues. This adds responseHeader to the list to prevent
these warnings.

* Pull condition fields in schema from KnownConditionFields var

This is to avoid having to keep the actual validation and the
documentation in sync in six places.