BUG/MAJOR: redactions: fix redactiontype key name
There's been an issue with redactions due to the key name being set to
redaction_type when the actual API key name is redactiontype [1]. This
commit updates it along with all of the references within examples to
be the correct name, redactiontype. Without this change the current
redaction logic does not appear to work at all.

[1] https://docs.fastly.com/signalsciences/api/#_corps__corpName__sites__siteName__redactions_post