BUG/MINOR: rules: suppress requestLogging default (#173)
requestLogging is an optional parameter that can be specified within
request rules. It defaults to "sampled", however, if you create a rule
without specifying it, it can cause terraform to mark a change as
happening even though one was not expected. e.g.

```
      + requestlogging  = "sampled"
```

In other situations, we could just set the default in the provider to
be "sampled". However, because requestLogging is not valid with all
rule types and at the time of this commit terraform does not allow you
to do pre-validation on multiple fields.

For this reason, we need to use `DiffSuppressFunc` to ignore the diff
if `requestLogging` was not specified but the API responded with
`sampled`.