BUG/MINOR: do not set default response code
`response_code` is only valid for for the `block` action. By setting a
default it forces it to be set for all action types including those that
it is not valid for. When `response_code` is empty on a blocking action
the API just defaults to using the site default (e.g. 406).

This commit removes the default from the Terraform provider for
`response_code` and also updates the documentation to indicate that it
is only valid for the `block` action type.