BUG/MINOR: lib: don't force response code to 406
In 4006a6c (BUG/MINOR: do not set default response code) the default to
406 was removed from rules without response codes. However, the provider
was still forcing a 406 default behind the scenes causing rules without
response codes to be recreated with 406 with every new plan apply.

This commit stops forcing the HTTP status code 406 on rules and fixes
the issue with rules that have not had changes happen from being
recreated.