MINOR: main: comment edge deployment by default (#196)
It's been reported that some users are starting with the default main.tf
This leads to having one of their sites erroneously configured as being
an Edge deployment. This commit comments out the edge deployment, which
allows it to be referenced but prevents their site from being configured
as an Edge deployment unless they want it to.