BUG/MEDIUM: rules: add redirect_url when flattening
It was reported that when using `redirect_url` changes were being
presented when applying a plan, even though there were not any changes
being made.

This was due to `redirect_url` not being reset during rule flattening
within `flattenRuleActions`.

This commit fixes that and adds `redirect_url` to the `actionmap` within
`flattenRuleActions`.