MEDIUM: site: add attack_threshold+immediate_block
This commit brings in the latest improvements that are in go-sigsci
0.1.19 (to be released), which allows for updating the attack thresholds
and/or setting the immediate block functionality.

attack_threshold can be set multiple times each for the various
intervals: 1, 10, 60 (in minutes).

immediate_block is a bool, which will instruct the WAF to immediately
block requests with attack signals.

This commit depends on go-sigsci#59 as well as a new version being
issued.