feat(sigsci_corp_user): Add corp user resource
This resource allows the provider to create, read, and delete a SigSci
Corp User. It also allows managing membership to sites and level of
access on initial creation. The upstream library (sigsci-go) does not
seem to support membership updates on a user. I would love to get some
feedback on how we should handle membership either here or in the
upstream module.