BUG/MEDIUM: edge: fix remapping of service to another workspace
It was recently identified by @BrooksCunningham that updating the
`site_short_name` on a `sigsci_edge_deployment_service` object resulted in an
error (`failed to attach service`). Successive runs of `terraform apply` would
result in a succes, however, ultimately it was not working as expected.

This update of the `site_short_name` is what is known and documented as a
"re-map" [1]. As the documentation shows, it requires first detaching the old
workspace from the service using a `DELETE` call and then attaching a new
workspace with a `PUT` call.

Previously, both Create & Update were using the same function, which prohibited
proper detachment.

This commit introduces a update specific function `updateEdgeDeploymentService`
and renames the existing `createOrUpdateEdgeDeploymentService` function to be
exclusively used with Create (`createEdgeDeploymentService`).

[1] https://docs.fastly.com/en/ngwaf/configuring-edge-waf-deployments-using-the-next-gen-waf-control-panel#re-mapping-a-fastly-cdn-service-to-a-new-site-workspace